pyproject.toml
[tool.poetry]
name = "pants-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.31"}
dependency-injector = "^4.38.0"
pydantic = "^1.9.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
docker = "^5.0.3"
black = "^22.1.0"
isort = "^5.10.1"
pytest = "^7.0.1"
docformatter = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100